Makes JSONDecoder configurable
